+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # SmartArray Changelog
 
+## [1.2.0] - 2024-10-31
+* Added sort() and sortBy($column) methods
+* Updated readme.md with examples
+* Misc code and other minor improvements
+
 ## [1.1.0] - 2024-10-28
 * Added unique() method
-* Misc code and other minor improvements.
+* Misc code and other minor improvements
 
 ## [1.0.0] - 2024-10-28
 * Initial release